Better playtime display and repacker help doc generation

+++ tools/process_new_game.ts
@@ -14,9 +14,15 @@ type ZipLevel = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
 interface CLIOptions {
   output: string
   zip_level?: ZipLevel
+  nopack: boolean
 
   [key: string]: string | number | boolean | ZipLevel | undefined
 }
+const obj_CLIOptions = {
+  output: "[./game_repacked] Output folder for the operation",
+  zip_level: "[7] Zip compressions level to use for packing",
+  nopack: "[false] Set to skip the packing stage of the repack"
+} as Record<string, string>;
 
 interface RepackCLIOptions extends CLIOptions {
   merge_data?: boolean
@@ -24,7 +30,19 @@ interface RepackCLIOptions extends CLIOptions {
   clear?: boolean
   nocleanup?: boolean
   dlc: string | true
+  skippackgame: boolean
 }
+
+const obj_RepackCLIOptions = {
+  ...obj_CLIOptions,
+  merge_data: "[false] Merge game data with remote data",
+  merge_info: "[false] Alias for merge_data",
+  clear: "[false] Only clear the output directory, no processing",
+  nocleanup: "[false] Skip the cleanup stage, will fully repack and leave unpacked files",
+  dlc: "[none] Set the dlc path for the game packing, use 'default' or true for ./proc/dlc",
+  skippackgame: "[false] Skips the packing of the game, but still generated all other files"
+} as Record<string, string>;
+
 interface DLCCLIOptions extends CLIOptions {
   input: string
   gameid: string
@@ -34,6 +52,19 @@ interface DLCCLIOptions extends CLIOptions {
   version?: string
   iter_id?: number
 }
+
+const obj_DLCCLIOptions = {
+  ...obj_CLIOptions,
+  input: "input folder for packaging",
+  gameid: "GameId for the parent game",
+  dlcid: "GamId for the DLC",
+  dlc_name: "Name of th DLC",
+  zip_name: "Name of the output zip file",
+  version: "[auto_id] The DLC version, if unset it will try and figure it out (unlikely)",
+  iter_id: "[auto_id] The DLC iteration id, if unset it will try and figure it out (unlikely)"
+} as Record<string, string>;
+
+
 type PathComponents = {
   dir: string
   file: string
@@ -624,14 +655,15 @@ async function repackGame(game_exe: string, output_dir: string, options: RepackC
   // Compress game data
   const zip_name = path.file.replace("setup_", "").replace(".exe", ".zip");
   const zip_output = output_dir + "/" + zip_name;
-  if(!options.nopack){
+  let dl_size = 0;
+  if(!options.nopack && !options.skippackgame){
     await compressFolder(game_folder, zip_output, options.zip_level);
+    dl_size = await getFileSize(zip_output);
   }
-  const dl_size = await getFileSize(zip_output);
   const slug = await getSlug(props, FORMAT);
   const version_data = await getVersion(game_exe, props, FORMAT);
   // Clean up left overs
-  if(!options.nocleanup){
+  if(!options.nocleanup && !options.nopack){
     removeDir(unpack_folder);
     removeDir(game_folder);
   }
@@ -676,10 +708,11 @@ async function repackDLC(dlc_exe: string, output_dir: string, options: RepackCLI
   // Compress game data
   const zip_name = path.file.replace("setup_", "").replace(".exe", ".zip");
   const zip_output = output_dir + "/" + zip_name;
+  let dl_size = 0;
   if(!options.nopack){
     await compressFolder(dlc_folder, zip_output, options.zip_level);
+    dl_size = await getFileSize(zip_output);
   }
-  const dl_size = await getFileSize(zip_output);
 
   // Make uninstall script
   const version_data = await getVersion(dlc_exe, props, FORMAT);
@@ -740,7 +773,9 @@ async function freshPackDLC(output_dir: string, options: DLCCLIOptions): Promise
   // Compress game data
   const zip_name = options.zip_name;
   const zip_output = output_dir + "/" + zip_name;
-  await compressFolder(dlc_folder, zip_output, options.zip_level);
+  if(!options.nopack){
+    await compressFolder(dlc_folder, zip_output, options.zip_level);
+  }
   const dl_size = await getFileSize(zip_output);
 
   // Make uninstall script
@@ -770,7 +805,7 @@ async function freshPackDLC(output_dir: string, options: DLCCLIOptions): Promise
   return dlc_frag;
 }
 
-async function main(game_exe: string, options_arr: string[]){
+async function packGame(game_exe: string, options_arr: string[]){
   const options = {
     output: "game_repacked",
     zip_level: 8,
@@ -778,6 +813,7 @@ async function main(game_exe: string, options_arr: string[]){
     merge_data: false,
     merge_info: false,
     nopack: false,
+    skippackgame: false,
     nocleanup: false,
     dlc: "NONE"
   } as RepackCLIOptions;
@@ -831,7 +867,7 @@ async function main(game_exe: string, options_arr: string[]){
     return;
   }
 
-  // Process the main game
+  // Process the packGame game
   let game = game_exe === "NONE" ? {} as GOG.RemoteGameData : await repackGame(game_exe, output_dir, options);
   const dlc = [] as GOG.RemoteGameDLCBuilding[];
   if(dlc_folder !== "NONE"){
@@ -858,6 +894,41 @@ async function main(game_exe: string, options_arr: string[]){
   }
 }
 
+async function batchPack(file_path: string, root_path: string){
+  if(!fs.existsSync(file_path)){
+    console.log("Could not find file_path [" + file_path + "]");
+    return;
+  }
+  if(!fs.existsSync(root_path)){
+    console.log("Could not find root_path [" + root_path + "]");
+    return;
+  }
+
+  type AutoJSON = {
+    folder: string
+    exe: string
+    dlc: boolean
+  }
+
+  const json = JSON.parse(fs.readFileSync(file_path).toString()) as AutoJSON[];
+
+  for(const game of json){
+    console.log("Processing game exe [" + game.exe + "] in [" + game.folder + "]");
+
+    const options = [
+      "output=game_repacked/out_" + game.folder,
+      "merge_data=true"
+    ];
+
+    if(game.dlc){
+      options.push("dlc=proc/" + game.folder + "/dlc");
+    }
+
+    await packGame(root_path + "/"  + game.folder + "/" + game.exe + ".exe", options);
+  }
+
+}
+
 async function packDLC(options_arr: string[]){
   const options = {
     output: "dlc_repacked",
@@ -866,7 +937,8 @@ async function packDLC(options_arr: string[]){
     zip_name: "REPLACE_ME",
     input: "REPLACE_ME",
     dlc_name: "REPLACE_ME",
-    zip_level: 8
+    zip_level: 8,
+    nopack: false
   } as DLCCLIOptions;
   if(options_arr){
     for(const part of options_arr){
@@ -910,12 +982,31 @@ async function packDLC(options_arr: string[]){
 
 /* global process */
 const args = process.argv.slice(2);
-if(args[0] === "packdlc"){
+if(args[0] === "help"){
+  console.log("Help: "
+  + "\n example: repack $exe_file [$options]"
+  + "\n - packdlc $options: Packs the given dlc for a game separate to the game ");
+
+  for(const key in obj_DLCCLIOptions){
+    console.log("    | - ", key, ":", obj_DLCCLIOptions[key]);
+  }
+
+  console.log(" - batch $batch_json $root_path: Batch process elements in the batch_json"
+  + "\n    | Json must be a list of objects"
+  + "\n    | objects must have folder:string, exe:string, dlc:boolean"
+  + "\n - [exe_file] [$repack_options]: Process the given EXE as a game to repack"
+  );
+  for(const key in obj_RepackCLIOptions){
+    console.log("    | - ", key, ":", obj_RepackCLIOptions[key]);
+  }
+}else if(args[0] === "packdlc"){
   packDLC(args.slice(1));
+}else if(args[0] === "batch"){
+  batchPack(args[1], args[2]);
 }else{
-  main(args[0], args.slice(1));
+  packGame(args[0], args.slice(1));
 }
 
 /* global exports */
-exports.main = main;
+exports.packGame = packGame;
 Object.defineProperty(exports, "__esModule", { value: true });
